move router service tests to their own file
